% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_create_table.R
\name{count_create_table}
\alias{count_create_table}
\title{Count occurrences of "create table" in a file}
\usage{
count_create_table(file_path)
}
\arguments{
\item{file_path}{Path to the file.}
}
\value{
Integer count of occurrences.
}
\description{
This function counts how many times the string "create table" appears in a file.
}
