% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsfirst.R
\name{dotsfirst}
\alias{dotsfirst}
\title{A sqlprep Function}
\usage{
dotsfirst(file_path)
}
\arguments{
\item{file_path}{Path to the data file}
}
\value{
A character vector with the converted first line
}
\description{
Convert "." to "_" in the first line of a data file
}
